[looker] Use original model names in deploy steps (Recidiviz/recidiviz-data#46267)
## Description of the change

In the initial roll out both the looker and prod projects were the same
as their respective model names and so i didn't distinguish between the
two. however, going forward we want to maintain the looker user
experience as much as we can. this means that the
`recidiviz-looker-staging` project should use model `recidiviz-staging`
and `recidiviz-looker` project should use model `recidiviz-123`. this
will ensure that all dashboard links remain the same since looker uses
the model name in the url not the project id.

in order to roll out this change we need to:
1. merge the corresponding [looker
changes](https://github.com/Recidiviz/looker/pull/835)
2. generate webhook token for `recidiviz-looker` project and update the
prod token in secret manager
3. merge this PR

we need to deploy both staging and prod at the same time because we need
to remove `recidiviz-staging` model from `recidiviz-looker` project
before deploying to staging, which will leave prod in a bad state. we
can try to piggy back this change onto someone else's CP deploy or merge
this on a friday and then manually deploy to looker production. looker
prod will be slightly out of sync for two days but i don't think it
would be a big deal since `recidiviz-looker` is currently a week+ stale.
then when the regular prod deploy went out on tuesday it would be a
no-op for looker.

4. remove `recidiviz-staging` model from `recidiviz-looker` project
5. reconfigure `recidiviz-looker-staging` to use only
`recidiviz-staging` model after the looker changes have been deploy to
staging
6. reconfigure `recidiviz-looker` to use only `recidiviz-123` model
after the looker changes have been deployed to prod
7. delete `recidiviz-looker-123` project
8. profit

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#36190

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9c8f523351eff82b88a1c74a63c997d88f68cda4