[preimport validations] Query postgres directly to get historical median row counts (Recidiviz/recidiviz-data#46473)
## Description of the change

- Updates `StableHistoricalRawDataCountsTableValidation` to use a bq
client in `us-east1` to query historical median row count from
`operations_v2_cloudsql_connection` (a view backed by a federated query
to operations cloud sql instance) and the normal bq client to get the
temp table row count.
- Updates pre import validations accordingly to accommodate the new
structure. this probably warrants a bigger refactor to make the
validation structure make more sense but it was becoming such a
undertaking that i didn't want to spend time on right now.

this change might make pre import validations a lot slower since we are
no longer running queries asynchronously. if so we can increase the
thread count.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#42453

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 02a6054b0e7ee83710b0381e3b3644a0deb25fb4