[US_ND] Update `MINIMUM_HOUSING_REFERRAL_QUERY` (Recidiviz/recidiviz-data#46481)
## Description of the change

We learned recently how to identify which facility a person was referred
to when a minimum housing request was made, regardless of whether or not
that request was subsequently approved.

This PR refactors the `MINIMUM_HOUSING_REFERRAL_QUERY` template so that
it reflects the finalized facility if there is one, and the requested
facility if there is not a reviewed and finalized entry.

This adds ~150 approved minimum housing referrals - this is because the
base raw data table is a more granular assessment table than we were
using before, so it catches more referrals than the previous query did.
I successfully loaded to a sandbox with prefix `elise_min`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Dan Hansen <dan@recidiviz.org>
Co-authored-by: Anna Geiduschek <ageiduschek@users.noreply.github.com>
GitOrigin-RevId: cf0a7f1d701202b2e28a7e463f2754bb571f4ae5