[Login Reminders] Add PA Supervisors (Recidiviz/recidiviz-data#46552)
## Description of the change

Add PA supervisors to email reminders, with their own specific text, not
the same as the supervision linestaff text - and add supervisor-specific
versions of the text for other states, to make it clearer that the
number is for people on the user's supervisees' caseloads. Also fix some
column names in the queries that were renamed.

Tested by clasp pushing to the real project and running the test
functions to write the emails to the spreadsheet, but I'll push and test
again after the parent PRs are rebased/merged!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8a01904d121e962e9665896c61dfcd3b70150b0c