[US_TN] Update StateSentenceStatusSnapshot ingest to be compatible with state_provided_start_date (Recidiviz/recidiviz-data#46557)
## Description of the change

I have tried many different configurations of ingest such that the
`SentenceEffectiveDate` ingested alongside statuses passes our checks.
None have worked. This PR changes our approach such that we ingest when
a sentence starts and when a sentence ends, but does not ingest other
statuses. This configuration passes our checks and maintains consistency
with effective date, which is what is needed for products.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes
[Recidiviz/recidiviz-data#44996](https://github.com/Recidiviz/recidiviz-data/issues/44996)
Closes
[Recidiviz/recidiviz-data#45513](https://github.com/Recidiviz/recidiviz-data/issues/45513)
Closes
[Recidiviz/recidiviz-data#44483](https://github.com/Recidiviz/recidiviz-data/issues/44483)
Closes
[Recidiviz/recidiviz-data#44222](https://github.com/Recidiviz/recidiviz-data/issues/44222)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0e94d2a59f44602a457152b62284a2facd074318