[US_AZ] Added historical kidnapping charge for exclusion (Recidiviz/recidiviz-data#46920)
## Description of the change

We got a report that someone was ineligible for TPR and not being
caught; after investigation, we found out that it is because of a
historical Kidnapping charge with an old statute code we were not
accounting for. With this change made, we find that there is a
change of one person, the exact person referenced in the PG ticket
[here](https://github.com/issues/assigned?issue=Recidiviz%7Crecidiviz-dashboards%7C8862).
Yahoo!

This change has been loaded to sandbox under `shalin`, all validations
are passing, and the change is, once again, one person is no longer
almost eligible for TPR. This is a .005 percent difference in
eligibility. One could say monumental, completely changes our tool

<img width="978" height="70" alt="Screenshot 2025-08-25 at 3 39 59 PM"
src="https://github.com/user-attachments/assets/587f18f7-6d14-460e-a037-3e20c88b2bf8"
/>

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes
[Recidiviz/recidiviz-data#8862](https://github.com/issues/assigned?issue=Recidiviz%7Crecidiviz-dashboards%7C8862)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 29771b31fbb83ac66f1102234194237e36d8d6ce