[raw data yaml writer] Implement updating list attributes (Recidiviz/recidiviz-data#46290)
## Description of the change

a `DirectIngestRawFIleConfig` and it's child classes might have a list
of python-native types or a list of child classes, which gets converted
to a list of dictionaries in the yaml updater.

in order to distinguish between a dict that has been updated and a net
new dictionary, we mark one of the class attributes with the
`LIST_IDENTIFIER_TAG` to indicate that that attribute can be used to
uniquely identify the object.

using that identifier, we compare list items and attempt to maintain the
original ordering as much as possible when items are added or deleted.
to do this, we first update the `ruamel_list` with any updated items,
deleted any deleted items, then we attempt to insert the net new items
after the previous item's location in the ruamel list and if we can't
find the previous item we insert at the end. sorry this is difficult to
describe but we essentially merge in any new items.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#46293

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 655f1b0b1fd5d45d751c9ded2b7b7c92a3ba63a2